requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmdstanpy==1.1.0
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
DateTime==5.2
decorator==5.1.1
ephem==4.1.4
fonttools==4.41.1
frozendict==2.3.8
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
holidays==0.29
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
LunarCalendar==0.0.9
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdtable==0.2
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.15.0
prophet==1.1.4
protobuf==4.23.4
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.15.1
PyMeeus==0.5.12
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==1.4.0
streamlit==1.25.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
streamlit-tags==1.2.8
ta==0.10.2
tenacity==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
webencodings==0.5.1
xgboost==1.7.6
yarl==1.9.2
yfinance==0.2.26
zipp==3.16.2
zope.interface==6.0
